// apps/server/src/utils/types.ts
export interface WebSocketData {
  id: string
  user: User
  channels: Set<string>
  socketId: string
  gameSocket: boolean
  developer: 'futex' | 'kickass' | 'nolimit' | 'netgame'
  connectedAt: number
  accessToken: string
  data?: Record<string, unknown>
}